Finish Get Post Put Delete Of news and comment

+++ Project2/src/main/java/se331/project2/DTO/News/NewsHomepageDTO.java
@@ -25,8 +25,8 @@ public class NewsHomepageDTO {
     Date publishedAt;
 
     NewsStatus status;
-    Long fakeCount;
-    Long notFakeCount;
+    int fakeCount;
+    int notFakeCount;
